feat: moved filter section to custom component
